requirements.txt
openai-whisper
pynput
sounddevice
scipy
pyperclip
pyobjc-framework-Quartz
pyobjc-framework-ApplicationServices
rumps
Pillow